Forgot to create a new branch for this but this includes the julia set problem for the homework.
To get this to work I needed to install freeglut3, not sure what is needed for
windows though. I also needed to add __device__ to the constructor.

It's still not working but I'll look at it tomorrow. It might be because I'm
doing this in an ssh session and don't have access to X.